Initial work for variable watching
